+++ src/components.d.ts
@@ -360,6 +360,9 @@ export namespace Components {
     'suffix': string;
     'value': number;
   }
+  interface ManifoldPerformance {
+    'ddLogs'?: any;
+  }
   interface ManifoldPlan {
     /**
     * URL-friendly slug (e.g. `"kitefin"`)
@@ -799,6 +802,12 @@ declare global {
     new (): HTMLManifoldNumberInputElement;
   };
 
+  interface HTMLManifoldPerformanceElement extends Components.ManifoldPerformance, HTMLStencilElement {}
+  var HTMLManifoldPerformanceElement: {
+    prototype: HTMLManifoldPerformanceElement;
+    new (): HTMLManifoldPerformanceElement;
+  };
+
   interface HTMLManifoldPlanElement extends Components.ManifoldPlan, HTMLStencilElement {}
   var HTMLManifoldPlanElement: {
     prototype: HTMLManifoldPlanElement;
@@ -1019,6 +1028,7 @@ declare global {
     'manifold-marketplace-grid': HTMLManifoldMarketplaceGridElement;
     'manifold-mock-resource': HTMLManifoldMockResourceElement;
     'manifold-number-input': HTMLManifoldNumberInputElement;
+    'manifold-performance': HTMLManifoldPerformanceElement;
     'manifold-plan': HTMLManifoldPlanElement;
     'manifold-plan-cost': HTMLManifoldPlanCostElement;
     'manifold-plan-details': HTMLManifoldPlanDetailsElement;
@@ -1064,7 +1074,7 @@ declare namespace LocalJSX {
   }
   interface ManifoldAuthToken extends JSXBase.HTMLAttributes<HTMLManifoldAuthTokenElement> {
     'oauthUrl'?: string;
-    'onManifold-token-receive'?: (event: CustomEvent<any>) => void;
+    'onManifold-token-receive'?: (event: CustomEvent<{ token: string }>) => void;
     /**
     * _(hidden)_ Passed by `<manifold-connection>`
     */
@@ -1410,6 +1420,9 @@ declare namespace LocalJSX {
     'suffix'?: string;
     'value'?: number;
   }
+  interface ManifoldPerformance extends JSXBase.HTMLAttributes<HTMLManifoldPerformanceElement> {
+    'ddLogs'?: any;
+  }
   interface ManifoldPlan extends JSXBase.HTMLAttributes<HTMLManifoldPlanElement> {
     /**
     * URL-friendly slug (e.g. `"kitefin"`)
@@ -1714,6 +1727,7 @@ declare namespace LocalJSX {
     'manifold-marketplace-grid': ManifoldMarketplaceGrid;
     'manifold-mock-resource': ManifoldMockResource;
     'manifold-number-input': ManifoldNumberInput;
+    'manifold-performance': ManifoldPerformance;
     'manifold-plan': ManifoldPlan;
     'manifold-plan-cost': ManifoldPlanCost;
     'manifold-plan-details': ManifoldPlanDetails;