+++ CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-
 - Added an events queue to `<manifold-performance>` to capture any performance events emitted before DataDog is ready.
+
+### Changed
+- Changed the docs fetch mocking to now wait for the real request duration. This duration is obtained at build time from the manifold APIs.
+
 
 ## [v0.5.8]
 
@@ -23,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Enforce standard height on product cards for more consistency
+- Replace auth token polling with an event based system
 
 ## [v0.5.7]